% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot fatal accidents by state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric, integer, or string specifying the
state of interest.}

\item{year}{A numeric, integer, or string specifying the year
of interest.}
}
\value{
A plot with points indicating fatal accidents within
   the state and year of interest.
}
\description{
A function to plot fatal accidents with a state from files
   that are made available by the US National Highway Traffic
   Safety Administration.
The function returns a plot as a side effect.
}
\section{Errors}{

The only years of data that are available are 2013, 2014,
   and 2015. If any other year is provided, an error is
   returned because the file does not exist.
If a number is provided that does not correspond to a state
   in the dataset, an error is returned.
If there are no accidents to plot in the state and year of
   interest, then an error message is returned.
}

\section{Imported functions}{

map() from maps package and points() from graphics package
}

\examples{
\dontrun{
fars_map_state(state.num = 6, year = 2015)
fars_map_state(6, 2015)
fars_map_state("6", "2015")
fars_map_state("6", "2014")
}

}
\seealso{
\href{https://cran.r-project.org/web/packages/maps/maps.pdf}{maps CRAN documentation} for more about maps package
}
