% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read fatality analysis reporting system (fars) data files}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the path to the file
to open.}
}
\value{
A tbl data frame from the dplyr package.
   There are no side effects.
}
\description{
A simple function to read in the compressed .csv.bz2
   files that are made available by the US National Highway Traffic
   Safety Administration.
The files are read in as tbl data frame objects from the dplyr
   package.
}
\section{Error}{

If the file path is not valid (e.g., file does not exist),
an error will result.
}

\section{Imported functions}{

read_csv() from readr and tbl_df() from dplyr
}

\examples{
fars_read("accident_2013.csv.bz2")

}
\seealso{
\href{http://readr.tidyverse.org/reference/read_delim.html}{read_csv} for more about reading in csv's using dplyr

 \href{https://www.rdocumentation.org/packages/dplyr/versions/0.5.0/topics/tbl_df}{tbl_df} for more about the tbl_df class
}
