% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in files using vector of year(s)}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of strings or numerics specifying the
years of data to open.}
}
\value{
A list of tbl data frames. There are no side effects.
}
\description{
A function to read in the compressed .csv.bz2
   files that are made available by the US National Highway
   Traffic Safety Administration. The user specifies a vector
   of years, and subsets of the data are read in to include
   only year and month columns.
The files are read in as tbl data frame objects from the dplyr
   package.
}
\section{Warning}{

The only years of data that are available are 2013, 2014,
   and 2015. If any other year is provided, a warning is given
   naming the invalid year.
}

\section{Imported functions}{

mutate() and select() from dplyr
}

\examples{
fars_read_years(c("2013","2014","2015"))
fars_read_years(c("2013","2014",2015))
fars_read_years(c("2013"))
fars_read_years(2013)
fars_read_years(c("2013","2014","2050"))

}
\seealso{
\href{https://cran.r-project.org/web/packages/dplyr/dplyr.pdf}{dplyr CRAN documentation} for more about dplyr
}
