% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize fars data files by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of strings or numerics specifying the
years of data to open.}
}
\value{
A tbl data frame with counts of observations from
   every month and year. There are no side effects.
}
\description{
A function to read in multiple years of files that are
   made available by the US National Highway Traffic
   Safety Administration.
The files are read in as a list, appended, and summarized
   using dplyr and tidyr functions.
}
\section{Warning}{

The only years of data that are available are 2013, 2014,
   and 2015. If any other year is provided, a warning is given
   naming the invalid year.
}

\section{Imported functions}{

bind_rows(), group_by(), and summarize() from dplyr and spread()
   from tidyr.
}

\examples{
\dontrun{
fars_summarize_years(c(2013,2050))
fars_summarize_years(2013)
fars_summarize_years(c("2013","2014","2015"))
}

}
\seealso{
\href{https://cran.r-project.org/web/packages/dplyr/dplyr.pdf}{dplyr CRAN documentation} for more about dplyr

 \href{hhttp://garrettgman.github.io/tidying/}{Data Science with R}
   to learn about data tidying
}
