% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Build filenames for fars data}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character, numeric, or logical object that is
converted to an integer.}
}
\value{
The name of a .csv.bz2 file
}
\description{
A simple function to build the name of compressed .csv.bz2
   files that are made available by the US National Highway
   Traffic Safety Administration.
}
\section{Warning}{

If the file path is a chacter string that can't be converted to an
   integer.
}

\section{Imported functions}{

None
}

\examples{
make_filename("2015")
make_filename(2015)

}
